Hold down for continuous cast + code fmt
